pyproject.toml
[tool.poetry]
name = "fastapi-sso"
version = "0.1.0"
description = ""
authors = ["anchellon <anshulmathew@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.114.1"}
authlib = {extras = ["client"], version = "^1.3.2"}
httpx = "^0.27.2"
itsdangerous = "^2.2.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
